LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/track-sizing/auto-repeat/intrinsic-auto-repeat/column-auto-repeat-auto-009-expected.txt


FAIL .masonry 1 assert_equals:
<div class="masonry" data-expected-width="300" data-expected-height="200">
  <div class="item" data-offset-x="75" data-offset-y="0" data-expected-width="150" data-expected-height="200"></div>
</div>
offsetLeft expected 75 but got 0
FAIL .masonry 2 assert_equals:
<div class="masonry" style="width: 200px; height: 100px;" data-expected-width="300" data-expected-height="200">
  <div class="item" data-offset-x="75" data-offset-y="0" data-expected-width="150" data-expected-height="200"></div>
</div>
offsetLeft expected 75 but got 0
FAIL .masonry 3 assert_equals:
<div class="masonry" style="width: min-content; height: min-content;" data-expected-width="300" data-expected-height="200">
  <div class="item" data-offset-x="75" data-offset-y="0" data-expected-width="150" data-expected-height="200"></div>
</div>
offsetLeft expected 75 but got 0
FAIL .masonry 4 assert_equals:
<div class="masonry" style="width: max-content; height: max-content;" data-expected-width="300" data-expected-height="200">
  <div class="item" data-offset-x="75" data-offset-y="0" data-expected-width="150" data-expected-height="200"></div>
</div>
offsetLeft expected 75 but got 0
FAIL .masonry 5 assert_equals:
<div class="masonry border" data-expected-width="320" data-expected-height="220">
  <div class="item" data-offset-x="75" data-offset-y="0" data-expected-width="150" data-expected-height="200"></div>
</div>
offsetLeft expected 75 but got 0
FAIL .masonry 6 assert_equals:
<div class="masonry border" style="width: 200px; height: 100px;" data-expected-width="320" data-expected-height="220">
  <div class="item" data-offset-x="75" data-offset-y="0" data-expected-width="150" data-expected-height="200"></div>
</div>
offsetLeft expected 75 but got 0
FAIL .masonry 7 assert_equals:
<div class="masonry border" style="width: min-content; height: min-content;" data-expected-width="320" data-expected-height="220">
  <div class="item" data-offset-x="75" data-offset-y="0" data-expected-width="150" data-expected-height="200"></div>
</div>
offsetLeft expected 75 but got 0
FAIL .masonry 8 assert_equals:
<div class="masonry border" style="width: max-content; height: max-content;" data-expected-width="320" data-expected-height="220">
  <div class="item" data-offset-x="75" data-offset-y="0" data-expected-width="150" data-expected-height="200"></div>
</div>
offsetLeft expected 75 but got 0
